RavenDB-11614 : move mixed-cluster-tests to a separate file, add interversion smuggler-tests
